backend/requirements.txt
amqp==1.4.9
annotated-types==0.5.0
anyjson==0.3.3
asgiref==3.5.2
astroid==2.12.12
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==1.7.0
backports.zoneinfo==0.2.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
dill==0.3.6
distlib==0.3.7
Django==4.1
django-celery==3.3.1
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.3
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.10.1
joblib==1.3.2
kombu==3.0.37
lazy-object-proxy==1.7.1
mccabe==0.7.0
numpy==1.24.4
pandas==2.0.3
Pillow==9.2.0
pipenv==2023.7.23
platformdirs==3.10.0
psycopg2-binary==2.9.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pylint==2.15.5
pyOpenSSL==23.1.1
python-dateutil==2.8.2
pytz==2022.5
scikit-learn==1.3.0
scipy==1.10.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.3
virtualenv-clone==0.5.7
wrapt==1.14.1
zope.interface==6.0
